Add extra abstraction for LineChartDataModel
