+++ appinventor/appengine/src/com/google/appinventor/client/editor/simple/components/MockAreaChartDataModel.java
@@ -2,7 +2,7 @@ package com.google.appinventor.client.editor.simple.components;
 
 import org.pepstock.charba.client.data.Data;
 
-public class MockAreaChartDataModel extends MockLineChartDataModel {
+public class MockAreaChartDataModel extends MockLineChartBaseDataModel {
     public MockAreaChartDataModel(Data chartData) {
         super(chartData);
     }
